catch watch chan closing
before the for and select would go into an infinite loop and spin as
fast as possible since <-w.ResultChan() will always succeed. Now when
the watch chan closes the for loop will exit. In order to also listen
for context cancel i added the go routine that will call w.Stop().
Checking if it is safe to call w.Stop() after the ResultChan has closed.